+++ packages/cmf/src/cmfConnect.js
@@ -26,7 +26,7 @@ import React from 'react';
 import hoistStatics from 'hoist-non-react-statics';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import { connect, useStore } from 'react-redux';
-import { v4 as uuidv4 } from 'uuid';
+import { randomUUID } from '@talend/utils';
 import actions from './actions';
 import actionCreator from './actionCreator';
 import component from './component';
@@ -235,7 +235,7 @@ export default function cmfConnect({
 		}
 
 		function CMFContainer(props, ref) {
-			const [instanceId] = React.useState(uuidv4());
+			const [instanceId] = React.useState(randomUUID());
 			const registry = React.useContext(RegistryContext);
 			const store = useStore();
 
@@ -334,7 +334,7 @@ export default function cmfConnect({
 		};
 		CMFContainer.WrappedComponent = WrappedComponent;
 		CMFContainer.getState = getState;
-
+		// eslint-disable-next-line @typescript-eslint/default-param-last
 		CMFContainer.setStateAction = function setStateAction(state, id = 'default', type) {
 			if (typeof state !== 'function') {
 				return getSetStateAction(state, id, type);