+++ packages/components/src/Breadcrumbs/Breadcrumbs.component.js
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
-import uuid from 'uuid';
 import { withTranslation } from 'react-i18next';
+import { randomUUID } from '@talend/utils';
 
 import theme from './Breadcrumbs.module.scss';
 import { Action, ActionDropdown } from '../Actions';
@@ -164,7 +164,7 @@ if (process.env.NODE_ENV !== 'production') {
 }
 
 BreadcrumbsComponent.defaultProps = {
-	id: uuid.v4(),
+	id: randomUUID(),
 	items: [],
 	maxItems: DEFAULT_MAX_ITEMS,
 	t: getDefaultT(),