+++ packages/design-system/src/components/InlineEditing/Primitives/InlineEditingPrimitive.tsx
@@ -8,8 +8,8 @@ import React, {
 	useEffect,
 	useState,
 } from 'react';
-import useKey from 'react-use/lib/useKey';
 import classnames from 'classnames';
+import keycode from 'keycode';
 import { useTranslation } from 'react-i18next';
 import { I18N_DOMAIN_DESIGN_SYSTEM } from '../../constants';
 import Form from '../../Form';
@@ -28,9 +28,15 @@ type ErrorInEditing =
 			description?: string;
 	  };
 
+export type OnEditEvent =
+	| React.MouseEvent<HTMLButtonElement>
+	| KeyboardEvent
+	| React.FormEvent<HTMLFormElement>
+	| React.KeyboardEvent;
+
 export type InlineEditingPrimitiveProps = {
 	loading?: boolean;
-	onEdit?: (event: React.MouseEvent<HTMLButtonElement> | KeyboardEvent, newValue: string) => void;
+	onEdit?: (event: OnEditEvent, newValue: string) => void;
 	onCancel?: () => void;
 	onToggle?: (isEditionMode: boolean) => void;
 	defaultValue?: string;
@@ -76,7 +82,7 @@ const InlineEditingPrimitive = forwardRef(
 			}
 		}, [hasError]);
 
-		const handleSubmit = (event: React.MouseEvent<HTMLButtonElement> | KeyboardEvent) => {
+		const handleSubmit = (event: OnEditEvent) => {
 			event.stopPropagation();
 			if (onEdit) {
 				const sentValue = value || '';
@@ -92,19 +98,6 @@ const InlineEditingPrimitive = forwardRef(
 			toggleEditionMode(false);
 			onCancel();
 		};
-
-		// Keyboard shortcuts
-		useKey('Escape', handleCancel, {}, [isEditing]);
-		useKey(
-			'Enter',
-			(event: KeyboardEvent): void => {
-				if (mode !== 'multi') {
-					handleSubmit(event);
-				}
-			},
-			{},
-			[isEditing, value],
-		);
 
 		const testId = `inlineediting.${mode === 'multi' ? 'textarea' : 'input'}`;
 
@@ -132,6 +125,7 @@ const InlineEditingPrimitive = forwardRef(
 
 		const sharedInputProps = {
 			'data-test': testId,
+			'data-testid': testId,
 			hideLabel: true,
 			hasError,
 			description,
@@ -142,10 +136,25 @@ const InlineEditingPrimitive = forwardRef(
 			onChange: (
 				event: React.ChangeEvent<HTMLInputElement> | React.ChangeEvent<HTMLTextAreaElement>,
 			): void => setValue(event.target.value),
+			// Keyboard shortcuts
+			onKeyDown: (event: React.KeyboardEvent) => {
+				if (event.keyCode === keycode.codes.enter && mode !== 'multi') {
+					// Enter
+					handleSubmit(event);
+				}
+				if (event.keyCode === keycode.codes.esc) {
+					handleCancel();
+				}
+			},
 		};
-
 		return (
-			<div {...rest} data-test="inlineediting" className={styles.inlineEditor} ref={ref}>
+			<div
+				{...rest}
+				data-test="inlineediting"
+				data-testid="inlineediting"
+				className={styles.inlineEditor}
+				ref={ref}
+			>
 				{isEditing ? (
 					<>
 						<div className={styles.inlineEditor__editor}>
@@ -168,6 +177,7 @@ const InlineEditingPrimitive = forwardRef(
 									<ButtonIcon
 										onClick={handleCancel}
 										icon="cross-filled"
+										data-testid="inlineediting.button.cancel"
 										data-test="inlineediting.button.cancel"
 										size="XS"
 									>
@@ -176,6 +186,7 @@ const InlineEditingPrimitive = forwardRef(
 									<ButtonIcon
 										onClick={handleSubmit}
 										icon="check-filled"
+										data-testid="inlineediting.button.submit"
 										data-test="inlineediting.button.submit"
 										size="XS"
 									>
@@ -199,6 +210,7 @@ const InlineEditingPrimitive = forwardRef(
 						<ValueComponent />
 						<span className={styles.inlineEditor__content__button}>
 							<ButtonIcon
+								data-testid="inlineediting.button.edit"
 								data-test="inlineediting.button.edit"
 								onClick={() => toggleEditionMode(true)}
 								aria-label={ariaLabel || label}