+++ packages/storybook/.storybook/docs/tokensDocHelpers/components/DefinitionList/DefinitionListBreakpoint.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import { Token } from '../../../../../src/tokens/types';
+import { TokensProps } from '../../TokensTypes';
+import { StackVertical } from '@talend/design-system';
+import DefinitionListItemBreakpoint from './DefinitionListItem/DefinitionListItemBreakpoint';
+
+const TokensDefinitionList = ({ tokens }: TokensProps) => {
+	return (
+		<StackVertical gap="S">
+			{tokens.map((token: Token, index: number) => (
+				<DefinitionListItemBreakpoint key={`${token.name}-${index}`} token={token} />
+			))}
+		</StackVertical>
+	);
+};
+
+export default TokensDefinitionList;