+++ tools/scripts-config-react-webpack/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @talend/scripts-config-react-webpack
 
+## 15.2.0
+
+### Minor Changes
+
+- 925cec02b: feat: add toggle to activate or not dynamic-cdn-plugin
+
+### Patch Changes
+
+- Updated dependencies [925cec02b]
+  - @talend/scripts-config-cdn@10.6.0
+
+## 15.1.2
+
+### Patch Changes
+
+- e83d88f83: fix: force postcss8 to avoid issue at build
+
+  we use autoprefixer 10 in this preset which require postcss8.
+  To avoid error where we could have postcss 7 let's force it to 8.
+
 ## 15.1.1
 
 ### Patch Changes